move image extract component into block
